+++ torchbearer/callbacks/__init__.py
@@ -60,6 +60,14 @@ Learning Rate Schedulers
         :members:
         :undoc-members:
 
+
+Learning Rate Finders
+------------------------------------
+
+..  automodule:: torchbearer.callbacks.lr_finder
+        :members:
+        :undoc-members:
+
 Weight Decay
 ------------------------------------